Rename StarCraftEnv, stub out basic tests
